FEAT :: Prgr_두원사이의정수쌍 풀이 수정(overflow 문제)

+++ src/main/kotlin/Prgr_두원사이의정수쌍.kt
@@ -6,8 +6,8 @@ class Prgr_두원사이의정수쌍 {
     fun solution(r1: Int, r2: Int): Long {
         var answer: Long = 0
         for (i in 1 until r2 + 1) {
-            val y2 = sqrt((r2 * r2 - i * i).toDouble()).toLong()
-            val y1 = ceil(sqrt((r1 * r1 - i * i).toDouble())).toLong()
+            val y2 = sqrt((r2.toDouble() * r2 - i.toDouble() * i)).toLong()
+            val y1 = ceil(sqrt((r1.toDouble() * r1 - i.toDouble() * i))).toLong()
             answer += y2 - y1 + 1
         }
         return (answer * 4)
